Add a function to convert sockaddr to string, and update raddr_to_str semantics
Add a function that does a clean conversion from a sockaddr to a string
and properly detects and handles truncation of the data.

Modify gensio_raddr_to_str() so that the return value of pos is where
it would have to been if buflen was large enough to hold the whole
output string.  This allows detection of the amount of data required for
the string, and detection of truncation.

Signed-off-by: Corey Minyard <cminyard@mvista.com>