Provide fd_ll access to read/write/except ready, write, and move tcp urgent
Let an fd_ll be able to get the ready operations, and give it access
to the write data coming from the gensio functions.  This will let
an fd_ll be able to do it's own read/write handling, so the SCTP can
handle multiple streams and represent them as channels.

As part of this, move the TCP urgent handling into the TCP code and
remove it from the generic ll_fd code.

Signed-off-by: Corey Minyard <cminyard@mvista.com>