Rework the UDP code
Greatly simplify the UDP code.  Consolidate the ndata allocation
function, rework the read disable code to be more reliable.

Also, get rid of the concept of resurrecting old gensios in the
closed state.  That could result in race conditions for the user,
if a UDP packet came in while the gensio was being closed.

Signed-off-by: Corey Minyard <cminyard@mvista.com>