Rework SSL authentication
Don't require a CA, allows the default one to be used.  Thus get
rid of "noCA" option.

Add a "clientauth" option to require that the server authenticate
the client.

Add a GENSIO_EVENT_PRECERT_VERIFY and GENSIO_ACC_EVENT_PRECERT_VERIFY
callback that is called from SSL before certification has started to
let the user modify the CA or otherwise reject the connection based
on the certificate.

Add a GENSIO_CONTROL_CERT_AUTH option to set the certificate
authority for a connection.

Signed-off-by: Corey Minyard <cminyard@mvista.com>