Big rework of defaults and options for serial devices
Make option keys case insensitive.

Convert a lot of things like LOCAL/-LOCAL to local[=true|false]
to make it more consistent with the rest of gensio.  Leave the old
style in for backwards compatibility, but it's deprecated.

Deprecate partiy, databits, etc, and just use the 9600n81 style.
Get rid of defaults for those.

Signed-off-by: Corey Minyard <cminyard@mvista.com>