Fix a number of issues found when documenting the Python interface
request_password() in the gensio_accepter had some issues with
the interface and the python state.

GENSIO_ACC_EVENT_POSTCERT_VERIFY was missing.  Add it into all
places that need it.

Remove disable from the gensio_accepter.  Python shouldn't need it.

Fix a few error reports with the wrong name.

Use "flush", not "sflush" as the rfc2217 callback because it can happen
on both the client and server.

Signed-off-by: Corey Minyard <cminyard@mvista.com>