base: Fix error handling from filter_ll_write()
Errors were not being handled from filter_ll_write().  This caused hangs
in openssl() where it would get into a situation where it couldn't
inject data into the lower layer but could not return a proper error.

Also, don't call the upper layer read handler on an error unless it has
enabled read.  This means that it has to defer the error in some cases.

Signed-off-by: Corey Minyard <cminyard@mvista.com>