serial: Allow the setting of arbitrary baud rates if termios2
If we have termios2, allow custom baud rates to be set if the custspeed
flag is passed in the serialdev gensio config.  If the hardware supports
it, then the speed can be set to any arbitrary integer.  Otherwise an
error is returned when setting the baud.

This only works for Linux at the moment.

Signed-off-by: Corey Minyard <cminyard@mvista.com>