gensio: Add some snprintf helper functions
Add a function to handle partial prints to buffers,
gensio_pos_snprintf().  You pass in the current position and the length,
if the position is past the length then nothing is done, just the number
of chars that would have been output returned.

Also add gensio_argv_snprintf() to printing argv strings and quoting
them properly.  Same rules as above when no space is left.

Also, in an unrelated change, modify gensio_raddr_to_str() to descend
down the children until it finds a gensio that works.

Signed-off-by: Corey Minyard <cminyard@mvista.com>