Require all accepter sockets to open, not just one
For some reason the open socket call would scan through the addresses
and try to open them, and if any of them opened it returned success.
I'm not sure why it was doing this, but it's not really ideal behavior,
you want it to get them all, but it was there for a reason that I can't
remember, and may not matter any more.

Anyway, getting ready for allowing zero to say "open using any dynamic
port" it will have to change.

Signed-off-by: Corey Minyard <cminyard@mvista.com>