Rework port zero handling for tcp, udp, and sctp
Setting the port to zero causes a non-assigned port to be used on some
OSes, and doesn't work on others.  Plus, if multiple addresses are given
to the accepter, even with that they might choose different port
numbers.

Modify the operation so that port zero work on all OSes as it should.
And force all the different addresses for an accepter to use the same
port.

This requires reworking some semi-internal interfaces dealing with
ports and socket opening.  Hopefully no external programs are using
those at the moment, but the changes resulted in quite a bit of
simplification, so they are worth it.

Signed-off-by: Corey Minyard <cminyard@mvista.com>