selector: Add an selector call to install a sigmask
Some programs may need to block using their own custom sigmask, so add a
way to do that.

This required reworking the old select-based selector code (when epoll
is not available) to use pselect instead.  This does lift the whole
"write to timeval" thing for handling racing signals and selects,
which was ugly in extreme.

Make the new sigmask waiter available in os funcs and waiters, too.
waiters were reworked and simplified a bit due to the new changes.

Signed-off-by: Corey Minyard <cminyard@mvista.com>