Fix error shutdown issues and bad NULL check
In some cases we would get a double free because a passed in child
gensio would be freed automatically.  Do an extra get on the child
to avoid this.

In another case the telnet could would double free the data because
freeing the filter frees the data.  Don't free the data on an err
if you freed the filter.

Also, the telnet code was checking the wrong value for NULL after an
alloc.

Signed-off-by: Corey Minyard <cminyard@mvista.com>