base: Fix some locking and usage issues
There was an issue that could cause the user handler to be called twice
with the same data in basen_ll_read(), fix that.

There was one spot with a missing enable check.

Add locking in basen_read_data_handler().  It may not be necessary, but
it's better safe here.

Signed-off-by: Corey Minyard <cminyard@mvista.com>