base: Don't automatically close a gensio on err if no callback set
There were races where the error would happen before the callback got
set, and the close would happen.

Instead assume that the callback will eventually be set and read
enabled, which will cause the read handler to report the error then.

Also, remove the recently added io check.  It should always be set, it
was another error that was causing it to be NULL.

Signed-off-by: Corey Minyard <cminyard@mvista.com>