Massive rework of things
It's hard to know where to start here.  But the basic state machine just
had a bunch of issues.  This adds a real state machine definition and
fixes everything to use it.

Lots of other things had to be reworked, debugging infrastructure added,
and too many other things.

The only API change is the list code now has some error checking, and
there is a new error.

Signed-off-by: Corey Minyard <cminyard@mvista.com>