mux: Fix a lockup issue and a bogus error report
If a channel gets a remote close request, don't return REMCLOSE in the
close response, otherwise you will get a read error on a channel you
requested close on.

Don't lock in muxc_alloc_channel_data, it's either called where a lock
is not necessary or called with the lock already held.

Plus remove the chan_check_read return which is no longer used.

Signed-off-by: Corey Minyard <cminyard@mvista.com>