stdio: Fix several problems
The accepter data structure wasn't being freed from an accepter.

Set the fd flags for either case (accepter or connector) where stdio is
self, and clean everything up properly.  Consolidate some of that code.

Fix error return handling on read, it wouldn't report properly in some
cases, resulting in the close error not being reported prpoerly.

Issue the close done properly for an accepter.

Allocate the gensio data structure initially for the accepter, and use
the enabled flag to know if it was set up.

Signed-off-by: Corey Minyard <cminyard@mvista.com>