relpkt: Rework ack handling
Have an ack send the next expected sequence, not the last received
message.  This remove a lot of +/- 1s in the code and simplifies things
a bit.

Also, if a ack is done for a packet that is awaiting transmission,
unschedule it for transmission.  This can happen on a retransmit
request.

Signed-off-by: Corey Minyard <cminyard@mvista.com>