relpkt: Rework resend_packets() handling
Pass in one greater than the last packet to send as the end packet.
This makes resending easy, because it detects automatically if first ==
end and doesn't send anything, and simplifies the logic a bit.

Otherwise if a resend is done and first == end, a bunch of packets are
resent.

Signed-off-by: Corey Minyard <cminyard@mvista.com>