Fix some library and include handling
Add OPENSSL_LIBS and friends to the LIBS and friends variable so they
don't have to be handled in each makefile, and so the gensio library is
linked directly against openssl.

This means Libs.private doesn't have to be done in libgensio.pc any
more.

Also move the base include directory handling to configure.as so it
doesn't have to be done in each makefile.

Signed-off-by: Corey Minyard <cminyard@mvista.com>