Get rid of unix-specific types in the public headers
The gensio header had addrinfo and sockaddr structs, and included
sys/socket.h, netdb.h, etc.  That's a bad idea, those things may not be
available on all systems.

Create a gensio_addrinfo structure and wrap it for now.  Move anything
dealing with sockaddr into gensio_osops.[ch] for now.

This is an API change, unfortunately.  Hopefully people are using the
more abstract interfaces to the code and won't run into this.

Signed-off-by: Corey Minyard <cminyard@mvista.com>