mux: Remove an incorrect write enable
There was a write enable right after the child open.  That's not
normally a problem, since it won't do anything until the child is open,
but it can race and happen after the child is open but before the open
call is complete.  Get rid of it, as it doesn't do anything useful,
anyway.

Signed-off-by: Corey Minyard <cminyard@mvista.com>