Get rid of the last os-specific types in the interface
Use unsigned long, not size_t, to define gensiods.

Create a new gensio_time type to replace struct timeval.

These are hopefully the last API changes that have to be made for a long
time, but are required to use this on systems besides *nix.

Pull the waiter code into gensio_selector and remove it's public
interface.

Signed-off-by: Corey Minyard <cminyard@mvista.com>