Rework network address scanning to allow a default
Set gensio_scan_network_port so that the default protocol can be passed
in to the protocol paramter return.  If it is set, use that protocol by
default if none is specified in the address string.

This way, you don't have to pass in the address type for local
addresses, You used to have to do things like:

  udp(laddr=udp,192.168.27.1,3000'),192.168.27.2,3000

Now you can leave of the "udp," from the laddr and it will default to
udp.

Signed-off-by: Corey Minyard <cminyard@mvista.com>