Rework IPv6 handling some
There were issues dealing with IPV6 and SCTP.  If you did the following
accepter:
  gensiot -a "sctp,::,3023"
and then connected with:
  gensiot "sctp,::1,3023"
it would disconnect in 2.5 seconds.  What was happening was the accepter
address would be IPv6 only, but the connecter address would allow V4
mapped addresses.  So the v4 addresses wouldn't get pings and the
disconnect happens.

Modify the operation of an address without "ipv6," explicitly in front of
it to all V4 mapped addresses.  And fix the connect side to set V6 only
on the socket if "ipv6,"is in front of it.

This means you cannot mix "ipv6," and "ipv6n4," addresses on the same
connection, and the default is now "ipv6n4"

This broke handling of things like "tcp,1234", so rework those to only
use IPv6 when it is mapped.

That, in turn, broke UDP.  Or, more accurately, expose a UDP bug, so
some work had to be done to handle IPv4 mapped addresses in IPv6 by
reworking address comparison to handle that.

Signed-off-by: Corey Minyard <cminyard@mvista.com>