Add GENSIO_CONTROL_RADDR, fix issues with GENSIO_CONTROL_LADDR
Add a control that is basically the same things as get_raddr_to_str().

Also make GENSIO_CONTROL_LADDR always take an input string in all cases.
It was in some cases ignoring the input and in other cases using it.
Make it consistenly use it.

Add printing the local and remote addresses to gensiot.

Signed-off-by: Corey Minyard <cminyard@mvista.com>