base: Add a way for a filter to say it can't take a write
If the upper layer enabled write but the filter was full and was waiting
for some remote flow control, the code would just spin waiting for a
write.  So add a way for the filter to say it is full, but it is
optional as most filter's don't need it.

Implement it for relpkt, which does need it.

Signed-off-by: Corey Minyard <cminyard@mvista.com>