base: Implement a way for the filter to say it has queued data
If a filter has queued data, but it's not ready to send (it's waiting
for acks from the other end), the state machine needs to go into write
drain state, not filter close state.  So implement a way to find that
out.  Add a relpkt way to handle that, too.

Also there were a few places where write drain state wasn't handled
properly.  Fix those.

Signed-off-by: Corey Minyard <cminyard@mvista.com>