ll_fd: Rework exception handling
There was a situation where a gensio had enabled it's child's write
enable, but not the read enable, due to data needing to be written.
The child failed, but this was not reported because exceptions don't
cause write waits to return, just read waits.

Rework it so that an exception will cause the write handle to be called
as well as the read handler.  This was somewhat complex because of TCP
urgent data; I had to rework that handling a little because I had to be
able to know if the exception was due to urgent data or due to a true
failure.  This had a few cascading effects into other bits of ll_fd
handling, unfortunately.

Signed-off-by: Corey Minyard <cminyard@mvista.com>