mux: Fix numerous bugs
Deliever "oob" to partial read reports that are oob.

Don't update received_unacked until after all the data is sent to the
user.  We unlock delivering the data, so if the received_unacked is
partially updated, that may be sent to the other end and the other end
may respond by sending more data before this end is ready.

If read is disabled in a partial read deivery, don't deliver any more
data.

Make sure all received data is delivered before sending a close.  That
way the proper acks can be sent before the close.  If the acks come
after the close, that is a protocol violation.

Signed-off-by: Corey Minyard <cminyard@mvista.com>