tests:oom: Switch relpkt tests to tcp
There appears to be a bug in Linux where sometimes a UDP socket can't
send to another UDP socket.  I've verified that packets are being sent,
all the addresses are correct, and you can open another socket and send
and it goes into the receiving socket.

Also add some docs on why.

Signed-off-by: Corey Minyard <cminyard@mvista.com>