Add a reuseaddr option for network gensios
It turns out the UDP issue was not a bug in Linux, but the fact that
SO_REUSEADDR was set on the UDP sockets and occasionally the two parts
of the test would choose the same socket port.

So add a "reuseaddr" parameter for network gensios, default TCP and
SCTP to be true, and default UDP to be false.

This required adding a class default for UDP at init time, that then
needed to be freed when cleaning out memory.

Signed-off-by: Corey Minyard <cminyard@mvista.com>