sergensio: When converting from a gensio, search children
If you get a sergensio event from a gensio, that might not be a serial
gensio, it may be that the serial gensio is one of the children.  So
when you do gensio_to_sergensio() on the received I/O, it might return
NULL.  So instead, search down the stack to see if any of the children
are sergensios and return the topmost one.

Since users of sergensio_to_gensio(), mostly for when getting user data,
expect it to return the same gensio as passed in to
gensio_to_sergensio(), the top level one the user deals with, make that
so.

This has some unfortuate side effects.  Some sergensio gensios were
using sergensio_to_gensio() to get their exact partner.  So add
sergensio_get_my_gensio() to get the exact partner and switch those.
It also means we have to track the associated gensio, which means
we need to set data dynamically, and thus need a lock in the gensio.

Signed-off-by: Corey Minyard <cminyard@mvista.com>

more