stdio: Rework read finish handling
Something was added to return if there was nothing to do, but that would
break things because there was code at the end of read finish handling
that needed to run even if there was nothing to deliver.  Rework the
code a bit to handle errors and data pending properly.

Signed-off-by: Corey Minyard <cminyard@mvista.com>