stdio: Fix locking in stdion_finish_read
It was, well, wrong.  It's hard to believe it worked as well as it did.
The changes for file I/O apparently shook things up enough to bring this
to the surface.

Anyway, stdion_finish_read() must be called with the lock held, and only
releases it for the callbacks.

Signed-off-by: Corey Minyard <cminyard@mvista.com>