gtlssh: Rework interactive password handling
There were a couple of issues if the first password (sent over the
certauth protocol) failed and we fell back to doing the password over
the connection:

The I/O for the password was being done on the certauth gensio, but it
should have bee done over the mux gensio.  Rework things to do the password
I/O over the proper gensio.

The sigwinch was coming in and messing up the first password (and thus
not getting the winch properly handled, either).  So add an oob command
back from gtlsshd to gtlssh to say that it is ready before sending the
initial sigwinch.

Signed-off-by: Corey Minyard <cminyard@mvista.com>