ssl: Handle some return errors in a special manner
In testing with injecting I/O errors, there comes one situation where
the SSL layer returns an error about getting data after a close is sent.
It appears to be a bug in the SSL code, or possibly some race condition
but I can't trace it down.

It's fairly harmless, anyway, but the tests fail if you get a protocol
error in this case, and it would confuse users.  So return an remote
close in this case.

This required reworking some error reporting because ssl_err had to be
fetched early.

Signed-off-by: Corey Minyard <cminyard@mvista.com>