sergensio: Round out the gensio accepter class handling
It was half-baked.  This allows you to set multiple classes for an
accepter, like you can for gensios.

This is primarily for telnet, so that users can tell if the telnet class
supports rfc2217.  It doesn't do anything else yet, but it's there for
completeness and may be required in the future.

Signed-off-by: Corey Minyard <cminyard@mvista.com>