perf: Always immediately disconnect on a try disconnect
The semantics on close are to not send any more data up if the user
calls gensio_close().  So we can't deliver the final data for perf.
This was resulting in the close not completing on a user close, since
the user disabled read and the gensio base code wasn't going to deliver
it, anyway.

It would be nice to deliver this data, but I'm not sure how.

Signed-off-by: Corey Minyard <cminyard@mvista.com>