selector: Keep track of each thread's wake time for waking threads
Stopping and starting timers can add quite a bit of overhead for
muliththreaded application, since adding a timer that is the shortest
requires all threads that are waiting to be signaled.

Instead, keep track of each thread's current wake time if it is waiting.
Don't signal the thread unless the earliest timer is earlier than the
thread's current wake time.

Signed-off-by: Corey Minyard <cminyard@mvista.com>