Promote the file descriptor for os funcs to an object
Having an int fd is just not very flexible, it make porting this to glib
or windows much harder.  Create a gensio_iod struct and use that instead
of an fd.  Unfortunately, this change propagated through a bunch of
code, and some gensios (stdio and serialdev) have to keep the fd and an
iod.

Signed-off-by: Corey Minyard <cminyard@mvista.com>