selector: Rework wait semantics to match expectations
The sel_select() functions and friends would always wait, even if they
handled something before the wait call.  This violates the expectation,
it's supposed to return if it handled anything.

So after running the runners and timers, only wait if nothing was
handled.

Also run the runners after the wait if it is done.  Otherwise there may
be runners sitting in the queue that were created as part of the fd
handling, and they might not get run because other threads waiting are
not woken for runners.  Not an issue for timers, as other threads are
woken if a timer is added to the top of the heap.

Signed-off-by: Corey Minyard <cminyard@mvista.com>