ipmisol: Fix a lock issue
sol_op_done() claimed the sol lock, but it was called with the sol lock
already held.  Modify to call sol_lock() in the callers that needed it.

Also don't do a sol_ref() in sol_start_op for operations that use a
runners.  The deferred op ref will be enough.

Signed-off-by: Corey Minyard <cminyard@mvista.com>