Fix IPV6/IPV4 multi-address handling
It appears that there is an issue with getaddrinfo(), if family ==
AF_UNSPEC it does not appear to return both IPV4 and IPV6 addresses.
I'm seeing this with localhost, which may be special, but if AF_UNSPEC
is set as the family and localhost as the host, it only returns one
address.  But you can set AF_INET6 and AF_INET4 and it will return the
proper one.

Rework the handling for AF_UNSPEC and no specified address type to scan
IPV6 and IPV4 addresses separately to work around the issue.

Signed-off-by: Corey Minyard <cminyard@mvista.com>