tcp: Add an option for tcpd handling and fix it
Allow tcpd to be enabled/disabled on a per-connection basis.  Also let
the option specify that "Access Denied" be written to the socket before
it is closed on refusal.

There was also a bug that caused a tcpd-refused connection to not be
closed properly, the file descriptor was leaked.  Fix that.

Signed-off-by: Corey Minyard <cminyard@mvista.com>