unix: Improve the efficiency of waiters
The waiter code was just waking every waiting thread whenever the count
was incremented.

Instead, put the count in the waiting thread struct and decrement the
count there, only waking the specific thread when the count gets to
zero.

This did bring up the issue of if the waiting thread returns when the
count is decremented some but not to zero, so in that situation re-add
whatever was decremented back to the waiter and pass it on to the next
threads.

Signed-off-by: Corey Minyard <cminyard@mvista.com>