udp: Fix the last UDP fix
The previous fix had an issue, disabling read in the wrong place could
cause the data to be freed while still in a read callback.

Only turn of in_read and deferred_read if the pending read is a deferred
one.  If it's in a callback, the return from the callback will finish
the close.

Signed-off-by: Corey Minyard <cminyard@mvista.com>