mux: Fix a refcount issue causing a free to never happen
A previous fix for this added a deref on shutdown, but that was
incorrect.  Remove that fix.

The basic problem that was happening was if you called gensio_free() on
a mux that was in a remove shutdown, it would never run through the
close code for mux channels, so it would not get the refcount
decremented from the close operation.

So, in the free operation, if we are in remote close state, do the
deref.

Fixes: ea5e5970ee mux: Fix a close/free issue
Signed-off-by: Corey Minyard <cminyard@mvista.com>