tools:gtlssh-keygen: Add an end-of-certificate detector to addallow -i
When reading data from stdin on addallow, detect the end of certificate
marker and shut down there instead of waiting for an end of file.

This is kind of a pain, but Windows has a habit of losing network data
when the program shuts down right after the final write/close.  This
way, we don't have to do any delays or anything like that, the addallows
will shut down when it has the whole certificate.

Signed-off-by: Corey Minyard <cminyard@mvista.com>