Windows: Fix numerous problems with incomplete read/writes
Really make sure a socket's write data is written before close by
looking for the close from the poll call.  This required a special
control be added, but so be it.

The previous change required reworking the stdsock code again for close.

When reading a oneway or twoway iod, don't return an error if there is
data pending.

Fix a wrong compare for the close socket call in all the os funcs.

Signed-off-by: Corey Minyard <cminyard@mvista.com>