win: Fix a race condition and use after free
Finishing freeing a waiter would release the lock after
calling the free finish, resulting in a use after free.

When shutting down a thread for an iod, do the done
setting and the wake in a single critical section to avoid
a race with the thread and a hang.

Also create a macro for locking the global lock to aid in
debugging issues.

Signed-off-by: Corey Minyard <cminyard@mvista.com>