addrinfo: Fix an issue with multiple addresses on connect socket
If you specify something like "localhost", you should get two addresses,
one for ipv4 and one for ipv6.  So don't do the retry code only on
listen sockets, do it on connect sockets to.  The net code will retry
on each address.

But not for SCTP yet, it needs some more work.

Signed-off-by: Corey Minyard <cminyard@mvista.com>