glib: Fix service to run something even if it would timeout now
If the timeout is zero, service wouldn't run anything necessarily.  At
least try to run something, and return the proper error if something was
run.

This was done by flagging if anything was done in all the handlers, and
using that in the service routine.  Also lower the priority of the
timeout in the wait routine.

Signed-off-by: Corey Minyard <cminyard@mvista.com>