oomtest: Fix stderr and shutdown issues
In order to get data from stderr from the subprocess, we have to wait
until the subprocess closes it, not close it as soon as we are done.
Close stdin to the subprocess and wait until the subprocess closes
stderr.

Also, free the os handler so it will report memory leaks.

Signed-off-by: Corey Minyard <cminyard@mvista.com>