swig: Fix a host of issues
Add a function to the waiter to run a service with a zero timeout and
return if it did anything, for clearing things out at the end.

Fix a host of mdns issues that came up with fixing the tests.

Fix newobject and refcounting when casting between gensio, sergensio,
and the accepter varieties.

Signed-off-by: Corey Minyard <cminyard@mvista.com>