mux: Rework channel refcounting a bit
Simplify things by having an open ref and an existance ref and only
deref the open ref when setting the instance to closed.

There was a situation where both channels send a new channel which would
result in an extra refcount for one of the channels.  In that situation,
lose a refcount.

Signed-off-by: Corey Minyard <cminyard@mvista.com>