mux: Fix a shutdown issue
Recent changes caused other bugs to show up, if you close the last
channel when a remote close (or other error) has been received, it would
try to re-close the child, resulting in an error return and bad things
happening.

Now don't call the child close if we have already done it.  This caused
another issue, the wrong close routine would be called, so hack that in.

Also add some missing refcount handling, and fix the mux_trace macro and
use it in one place.

Signed-off-by: Corey Minyard <cminyard@mvista.com>