tools: Rework gensiotool to ready for multi-connection threading
Remove the just-added fork thing.  fork() doesn't work on Windows, and
has performance issues.  Instead, rework the code to allow multiple
simultaneous connections.

Modify the way connections are created and destroyed to consolidate all
the operations and not do everything at exit time, but clean up and wake
when the cleanup is complete.

Rename some data structures to more sensible names, and move global data
in to a data structure for a cleaner design.

Signed-off-by: Corey Minyard <cminyard@mvista.com>