os_funcs: Add a reload and terminate handler
Add a way so users can register for events requesting termination and
events requestion a reload of data.  They come in as normal gensio
callbacks.

This required a tighter tying between the os handler and the proc data,
a proc data pointer is now kept in the os handler and used for waiting.

Signed-off-by: Corey Minyard <cminyard@mvista.com>