serialdev: Allow serial close timeouts to be set
The serial ports were set so that if data was pending and no data was
sent in .5 seconds, close the port.

This is not exactly ideal in all situations.  So allow an overall
timeout to be set, and a timeout when no data is sent for a period of
time.

Signed-off-by: Corey Minyard <cminyard@mvista.com>