mux: Increase the default buffer sizes
The defaults, 1024, were just inadequate.  For reasonable performance
the read buffer must be much larger than the write buffer, and the write
buffer needs to be bigger.

This lets multiple outstanding packets be in progress at the same time,
which wasn't possible with the previous numbers.

These are not high-performance numbers, but these can be tuned.

Signed-off-by: Corey Minyard <cminyard@mvista.com>