Modify oob handling to be disabled by default
Before, oob data delivery was enabled and could not be disabled.
However, many apps don't care about oob data, and that make it
inconvenient for each of them to handle it.  So disable it be default
and add a control to enable it.

Adjust the various users (telnet, gtlssh[d]) to enable it as necessary.
This required adding a base operation where a filter can set the child
controls.

Signed-off-by: Corey Minyard <cminyard@mvista.com>