Reject new channel on error return from new channel callback
You can to handle the return code from new channel, the upper layer may
return GE_NOSUP if it doesn't support it, and you have to do the right
thing in that case.

The mux changes uncovered some bugs in there, some refcount rework had
to be done to handle this, and free/close calls in the handler had to
have some special handling to avoid bad situations.

Signed-off-by: Corey Minyard <cminyard@mvista.com>