c++: Rework Os_Funcs logging and construction
The logging stuff was kind of a mess, and hard for integration with
SWIG.  Rework it to be class-based (more C++ like).

Also rework construction of Os_Funcs a bit.  There were places missing
reference settings and that was causing issues.  Instead, don't have a
constructor with no parameters and use the initialize method to set
everthing up.

Signed-off-by: Corey Minyard <cminyard@mvista.com>