Check return values from read and write ready events
Users should be returning 0 from these already, so it should be ok.  But
when handling errors and exceptions from callbacks into other code, we
need some way to handle exceptions and other errors of that sort.  If
they don't get handled properly, the program could go into an infinite
loop.  So when an error gets returned, at least stop calling the
callback event, and in most cases error out the connection.

Signed-off-by: Corey Minyard <cminyard@mvista.com>