c++: Remove Gensio and Accepter from callbacks
It's kind of nice to have that, but not really necessary.

It causes all kinds of problems when doing wrapping in other languages.
In different langauges converting from a class to the target language
class can be difficult.  In some cases you would have to have lookup
maps or something.

So just get rid of it.  It simplifies things greatly.

Signed-off-by: Corey Minyard <cminyard@mvista.com>