selector: Fix a problem with runners not running
If a runner is run from an outside thread (one that won't wait inside
the gensio library), it doesn't wake up anything else to handle the
runner, so it won't get run until something wakes up or calls back into
the library.

To fix this, send a signal to the first waiter.  It will handle the
runner upon waking.

Signed-off-by: Corey Minyard <cminyard@mvista.com>