selector: Remove the current waiter before re-applying the count
If a waiter returns an error, any count it received is applied to the
next waiter, making it indepotent if it returns an error.  The code,
however, didn't remove the first waiter before applying the count, so
the count would be re-applied to the first waiter.

Move the list removal to before re-applying the count.

Signed-off-by: Corey Minyard <cminyard@mvista.com>