tests:oomtest: Add .allow_no_err_on_trig to all tests
There are situations in all tests in a multi-threaded system that can
result in an error being triggered and it not causing the program to
fail.  For instance, a read can fail, but if the gensio is being closed,
that failure will not cause an error.  It's just too hard to get it
to work properly with that setting disabled.

Signed-off-by: Corey Minyard <cminyard@mvista.com>