serialdev: Fix serial break issues
Don't fail the connection if turning the break off fails.  This happens
on some USB and bluetooth devices because they don't implement it, but
there's no way to tell ahead of time if they will fail.  The failure
is basically "I don't support this".  Even though break is a pretty
fundamental thing for serial ports.  This is a regression, but the
rationale was not recorded in the code before.

Also, use the right value for the set break operation.

Signed-off-by: Corey Minyard <cminyard@mvista.com>