sergensio: Fix the class propagation
When doing sergensio calls, it was passing in the lowest-level sergensio
structure to the callback, which may not be the one the user is dealing
with since gensios may be stacked on top of it.

So catch each call at each level so you can pass back the proper
sergensio structure to the user.

Signed-off-by: Corey Minyard <cminyard@mvista.com>