Always free the gensio data first when freeing a gensio
Calling gensio_data_free() can cause function calls into the gensio if
the gensio is set up for synchronous I/O.  In some cases that could
result in freed data being used and thus crashes.

Move the call to gensio_data_free() to the top of the free function in
all gensios to avoid this issue.

Signed-off-by: Corey Minyard <cminyard@mvista.com>