certauth: Add support for ed25519 keys
ed25519 keys can't use EVP_SignInit and friends.  It turns out
EVP_SignInit and friends is the wrong thing to use, anyway, since they
don't hash the data first, per the EVP_PKEY_sign() man page.

So a version 4 protocol is added, and that uses EVP_DigestSign, which
will work with ed25519 keys if you do the sign and verify in one
chunk.  Thus the data to sign/verify has to be allocated and all the
chunks copied into it.  Sigh.

And ed25519 doesn't support a passed-in digest, so adjustments had to be
made for that.

Anyway, with this change, we can support ed25519.

Signed-off-by: Corey Minyard <cminyard@mvista.com>