tools:gtlsshd: Rework login and pam handling
Since we are using pam, login is not required, directly execute the
shell in a pty.

pam_open_session() needed to be called, along with pam_close_session().
pam_close_session() cleans up credentials, so no need to clean those up
ourselves.

nologin is handled by pam, don't do it in gtlssh.

Fix the handling of interactive logins.  Only generate login output if
we are logging in interactively.

Signed-off-by: Corey Minyard <cminyard@mvista.com>