mux: Fix an issue with I/O in the new connection handler
If a write is done in a new connection handler, the data may get there
before the new connection response, which is a protocol violation.  So
in that case go ahead and send the new connection response immediately.

However, we still have to handle an error return from the new connection
response.  In the above case, just do a normal close.

Signed-off-by: Corey Minyard <cminyard@mvista.com>