osops: Fix handling of ptys in an impersonation context
If impersonating an lower-privilege user, you cannot use CreatePipe
because the default security context for CreatePipe is inaccessible
to the lower privilege user.

So return to the normal security context while creating the pty and then
go back to the original security context after.

Also fix a call to OpenThreadToken, it must use the process context, not
the thread context, for the same reason.

Signed-off-by: Corey Minyard <cminyard@mvista.com>