pty: Fix a shutdown issue where the exit code couldn't be fetched
On Windows, closing the low-level PTY would result in an error because
it would shut down the PTY, close the child process handle as part of
that, and then check the child process error code on the child that is
now gone.

Add a stop control to PTY IODs so they can be stopped without closing
the child process.

Signed-off-by: Corey Minyard <cminyard@mvista.com>