script: Add a script filter gensio
At startup, this will execute an external program that is connected to
the child gensio of this gensio.  stdin/stdout of the external program
will be routed to/from the child gensio.

With this, you can write a script to do any setup on a connection.

When the script exits, this gensio will go into passthrough mode and
signal that it is open.

Signed-off-by: Corey Minyard <cminyard@mvista.com>